frontend/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pymupdf>=1.23.0 